Modernize project tooling and bump to 2.0.0
Sweeping modernization of the project's toolchain and public API
surface. See CHANGELOG.md for the detailed 2.0.0 release notes.

Highlights:

- Migrate from setup.py to pyproject.toml; adopt uv + uv_build.
  Commit uv.lock. Switch to src/ layout.
- Drop Python < 3.10; support 3.10 through 3.14; uv_build wheel
  ships py.typed for PEP 561 consumers.
- Adopt ruff (lint + format) and ty (types); remove black/flake8.
- pytest + pytest-cov with a 75% coverage floor.
- Modern typing: Literal aliases (OptionKind, InFormat, RateFreq,
  RfFreq, EwmParam), EwmParams TypedDict, RegressionResult Protocol.
- Pandas 2.x API compatibility fixes (fillna method=, delim_whitespace,
  skip_footer, is_all_dates, legacy M/Q/A freq codes, Index.FreqGroup
  enum lookup).
- NumPy 2.x compatibility (drop numpy.lib.nanfunctions imports).
- datasets.py: lazy pandas_datareader proxy to tolerate Py 3.12+
  distutils removal. load_13f now requires a user_agent per SEC EDGAR
  policy. Remove load_factors/load_shiller/load_retaildata (upstream
  sources drifted beyond repair).
- CI: SHA-pinned matrix workflow (3.10-3.14) with lint/test/build.
- publish.yml: tag-triggered OIDC trusted publishing + Sigstore +
  SBOM + GitHub Release (pinned SHAs, least-privilege per job).
- Dev ergonomics: Taskfile.yml, .pre-commit-config.yaml (prek-
  compatible), .editorconfig, Dependabot for github-actions + uv,
  SPDX license field.
- CHANGELOG.md follows Keep a Changelog.
- CLAUDE.md and a lint-clean README rewrite (markdownlint-cli2).
