+++ .editorconfig
@@ -1,0 +1,25 @@
+# EditorConfig — https://editorconfig.org
+# Python project defaults; aligned with ruff's formatter (line-length 88).
+root = true
+
+[*]
+charset = utf-8
+end_of_line = lf
+insert_final_newline = true
+trim_trailing_whitespace = true
+indent_style = space
+indent_size = 4
+max_line_length = 88
+
+[*.{py,pyi}]
+indent_style = space
+indent_size = 4
+
+[*.{toml,yml,yaml,json}]
+indent_size = 2
+
+[*.md]
+trim_trailing_whitespace = false
+
+[Makefile]
+indent_style = tab
